connected pages
