+++ app/src/main/res/layout/fragment_pin_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".pinList">
+    tools:context=".pages.pinList.pinListFragment">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
